Done controllers Product, Categories and orders

+++ src/main/java/edu/unapec/shoppingorders/repositories/ProductRepository.java
@@ -2,5 +2,5 @@ package edu.unapec.shoppingorders.repositories;
 
 import edu.unapec.shoppingorders.models.Product;
 
-public interface ProductRepository extends Repository<Product, Integer> {
+public interface ProductRepository extends Repository<Product, Long> {
 }
